+++ scripts/trunk_gh.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-if [[ -z "$TRAVIS" ]]; then
-	echo "Script is only to be run by Travis CI" 1>&2
+if [[ -z "$GITHUB_WORKFLOW" ]]; then
+	echo "Script is only to be run by GitHub Actions" 1>&2
 	exit 1
 fi
 
@@ -10,31 +10,20 @@ if [[ -z "$WP_PASSWORD" ]]; then
 	exit 1
 fi
 
-if [[ -z "$TRAVIS_BRANCH" || "$TRAVIS_BRANCH" != "production" ]]; then
-	echo "Build branch is required and must be a release-tag" 1>&2
+if [[ -z "$GITHUB_REF" || "$GITHUB_REF" == "refs/heads/production" ]]; then
+	echo "Build branch is required and must not be production" 1>&2
 	exit 0
 fi
 
-
 PLUGIN="eduadmin-booking"
-PROJECT_ROOT=$TRAVIS_BUILD_DIR
+PROJECT_ROOT=$GITHUB_WORKSPACE
 VERSION="$(cat $PROJECT_ROOT/eduadmin.php | grep Version: | head -1 | cut -d: -f2 | tr -d '[[:space:]]')"
 
 echo "Version: $VERSION of $PLUGIN"
 
-# Check if the tag exists for the version we are building
-TAG=$(svn ls "https://plugins.svn.wordpress.org/$PLUGIN/tags/$VERSION")
-error=$?
-if [ $error == 0 ]; then
-    # Tag exists, don't deploy
-    echo "Tag already exists for version $VERSION, aborting deployment"
-    exit 1
-fi
-
 # Remove files not needed in plugin for deployment
 rm -f $PROJECT_ROOT/composer.json
 rm -f $PROJECT_ROOT/.scrutinizer.yml
-rm -f $PROJECT_ROOT/.travis.yml
 rm -f $PROJECT_ROOT/CONTRIBUTING.md
 rm -f $PROJECT_ROOT/LICENSE.md
 rm -f $PROJECT_ROOT/phpunit.xml
@@ -42,6 +31,8 @@ rm -f $PROJECT_ROOT/.gitignore
 rm -f $PROJECT_ROOT/Gulpfile.js
 rm -f $PROJECT_ROOT/yarn.lock
 rm -f $PROJECT_ROOT/package.json
+rm -f $PROJECT_ROOT/readme.md
+rm -fR $PROJECT_ROOT/.github
 rm -fR $PROJECT_ROOT/scripts
 rm -fR $PROJECT_ROOT/tests
 rm -fR $PROJECT_ROOT/.git
@@ -55,6 +46,8 @@ rm -fR $PROJECT_ROOT/vendor
 rm -fR $PROJECT_ROOT/bin
 rm -fR $PROJECT_ROOT/node_modules
 rm -fR $PROJECT_ROOT/src
+rm -fR $PROJECT_ROOT/docs
+rm -fR $PROJECT_ROOT/website
 
 # Make sure we are in the project root
 cd $PROJECT_ROOT
@@ -78,28 +71,13 @@ svn co -q "http://svn.wp-plugins.org/$PLUGIN" svn
 # Copy our new version of the plugin into trunk
 rsync -r -p -v --delete-before $PROJECT_ROOT/* svn/trunk
 
-# Add new version tag
-mkdir svn/tags/$VERSION
-rsync -r -p -v --delete-before $PROJECT_ROOT/* svn/tags/$VERSION
-
-# Add new files to SVN
 svn stat svn | grep '^?' | awk '{print $2}' | xargs -I x svn add x@
 # Remove deleted files from SVN
 svn stat svn | grep '^!' | awk '{print $2}' | xargs -I x svn rm --force x@
 svn stat svn
 
 # Commit to SVN
-svn ci --no-auth-cache --username $WP_USERNAME --password $WP_PASSWORD svn -m "Deploy version $VERSION"
+svn ci --no-auth-cache --username $WP_USERNAME --password $WP_PASSWORD svn -m "Committing changes for $VERSION"
 
 # Remove SVN temp dir
 rm -fR svn
-
-curl -X POST \
--H 'Content-type: application/json' \
---data '{"username": "Travis CI", "channel":"#wordpress-eduadmin", "icon_url": "https://a.slack-edge.com/0180/img/services/travis_48.png","text": "EduAdmin Booking plugin version '"$VERSION"' deployed to <https://sv.wordpress.org/plugins/eduadmin-booking/|wp.org> :tada:"}' \
-$SLACK_HOOKURL
-
-#curl -X POST \
-#-H 'Content-type: application/json' \
-#--data '{ "fromUser": { "username": "TravisCI", "displayName": "Travis CI", "avatarUrlSmall": "https://a.slack-edge.com/0180/img/services/travis_48.png", "avatarUrlMedium": "https://a.slack-edge.com/0180/img/services/travis_48.png" }, "text": "EduAdmin Booking plugin version '"$VERSION"' deployed to [wp.org](https://sv.wordpress.org/plugins/eduadmin-booking/) :tada:" }' \
-#$GITTER_API